SeawaterPolynomials.jl

This package provides approximations to the Boussinesq equation of state for seawater expressed as polynomial functions of conservative temperature, absolute salinity, and geopotential height.

Computationally efficient polynomial approximations to the Boussinesq seawater equation of state are crucial components of ocean modeling software.

The Seawater Boussinesq approximation

In the seawater Boussinesq approximation, the density of seawater is expanded around a constant reference value ρᵣ,

ρ = ρᵣ + ρ′(Θ, Sᴬ, Z)

where the anomaly, ρ′, is a function of conservative temperature Θ, absolute salinity Sᴬ, and geopotential height Z. One choice for ρᵣ is the average density at the surface of the world ocean, ρᵣ = 1024.6 kg m⁻³, according to Roquet et al. (2015).

The TEOS-10 standard

The Thermodynamic Equation of SeaWater (TEOS-10) is a Gibbs function formulation of seawater themodynamics.

The error between the polynomials implemented in this package and the TEOS-10 is minimized for current 'climatological' ocean distributions of conservative temperature and absolute salinity. For more information, see

Roquet et al., "Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard", Ocean Modelling (2015)

Roquet et al., "Defining a Simplified Yet “Realistic” Equation of State for Seawater", Journal of Physical Oceanography (2015)

Related packages

GibbsSeaWater.jl

References

Roquet et al., "Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard", Ocean Modelling (2015)

Roquet et al., "Defining a Simplified Yet “Realistic” Equation of State for Seawater", Journal of Physical Oceanography (2015)

Young, W. R., "Dynamic Enthalpy, Conservative Temperature, and the Seawater Boussinesq Approximation", Journal of Physical Oceanography (2010)

API

SeawaterPolynomials.AbstractSeawaterPolynomial — Type

AbstractSeawaterPolynomial

Abstract type for Boussinesq equations of state expressed as polynomial functions of conservative temperature, absolute salinity, and geopotential depth.

SeawaterPolynomials.haline_contraction — Function

haline_contraction(Θ, Sᴬ, Z, equation_of_state)

Returns the haline contraction coefficient,

$β = ∂ρ / ∂Sᴬ$

for a seawater parcel with absolute salinity Sᴬ and at fixed conservative tempertuare Θ and geopotential height Z, using the Boussinesq equation_of_state. The haline contraction coefficient measures how much seawater density changes when absolute salinity is changed. 'Haline contraction' is so named because, due to sign convention, positive values reflect increasing seawater density with increasing absolute salinity, and thus a slight `contraction' of oceanic fluid parcels.

The geopotential height Z(x, y) is defined such that Z(x, y) = 0 at sea level and decreases downwards to negative values, towards the bottom of the ocean.

SeawaterPolynomials.thermal_expansion — Function

thermal_expansion(Θ, Sᴬ, Z, equation_of_state)

Returns the thermal expansion coefficient,

$α = - ∂ρ / ∂Θ$

for a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, and using the Boussinesq equation_of_state. The thermal expansion coefficient measures how much seawater density changes when conservative temperature is changed. 'Thermal expansion' is so named because, due to sign convention, positive values reflect decreasing seawater density with increasing conservative temperature, and thus an 'expansion' of oceanic fluid parcels. In many, but not all conditions in Earth's ocean (at temperatures greater than 4ᵒC in freshwater), the thermal expansion coefficient is positive.

The geopotential height Z(x, y) is defined such that Z(x, y) = 0 at sea level and decreases downwards to negative values, towards the bottom of the ocean.

SeawaterPolynomials.ρ — Method

ρ(Θ, Sᴬ, Z, equation_of_state)

Returns the total density of a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, using the Boussinesq equation_of_state.

The geopotential height Z(x, y) is defined such that Z(x, y) = 0 at sea level and decreases downwards to negative values, towards the bottom of the ocean.

This function aliases total_density.

SeawaterPolynomials.ρ′ — Function

ρ′(Θ, Sᴬ, Z, equation_of_state)

Returns the total density of a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, using the Boussinesq equation_of_state.

The geopotential height Z(x, y) is defined such that Z(x, y) = 0 at sea level and decreases downwards to negative values, towards the bottom of the ocean.

The function aliases density_anomaly.