API

SeawaterPolynomials

SeawaterPolynomials.AbstractSeawaterPolynomial — Type

AbstractSeawaterPolynomial

Abstract type for Boussinesq equations of state expressed as polynomial functions of conservative temperature, absolute salinity, and geopotential depth.

SeawaterPolynomials.haline_contraction — Function

haline_contraction(Θ, Sᴬ, Z, equation_of_state)

Returns the haline contraction coefficient,

\[β = ∂ρ / ∂Sᴬ ,\]

for a seawater parcel with absolute salinity Sᴬ and at fixed conservative tempertuare Θ and geopotential height Z, using the Boussinesq equation_of_state. The haline contraction coefficient measures how much seawater density changes when absolute salinity is changed. 'Haline contraction' is so named because, due to sign convention, positive values reflect increasing seawater density with increasing absolute salinity, and thus a slight `contraction' of oceanic fluid parcels.

The geopotential height is defined such that $Z(x, y) = 0$ at sea level and decreases downwards to negative values, towards the bottom of the ocean.

SeawaterPolynomials.thermal_expansion — Function

thermal_expansion(Θ, Sᴬ, Z, equation_of_state)

Returns the thermal expansion coefficient,

math α = - ∂ρ / ∂Θ ,`

for a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, and using the Boussinesq equation_of_state. The thermal expansion coefficient measures how much seawater density changes when conservative temperature is changed. 'Thermal expansion' is so named because, due to sign convention, positive values reflect decreasing seawater density with increasing conservative temperature, and thus an 'expansion' of oceanic fluid parcels. In many, but not all conditions in Earth's ocean (at temperatures greater than 4ᵒC in freshwater), the thermal expansion coefficient is positive.

The geopotential height is defined such that $Z(x, y) = 0$ at sea level and decreases downwards to negative values, towards the bottom of the ocean.

SeawaterPolynomials.ρ — Method

ρ(Θ, Sᴬ, Z, equation_of_state)

Returns the total density of a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, using the Boussinesq equation_of_state.

The geopotential height is defined such that $Z(x, y) = 0$ at sea level and decreases downwards to negative values, towards the bottom of the ocean.

This function aliases total_density.

SeawaterPolynomials.ρ′ — Function

ρ′(Θ, Sᴬ, Z, equation_of_state)

Returns the total density of a seawater parcel with conservative tempertuare Θ, absolute salinity Sᴬ, at the geopotential height Z, using the Boussinesq equation_of_state.

The geopotential height is defined such that $Z(x, y) = 0$ at sea level and decreases downwards to negative values, towards the bottom of the ocean.

The function aliases density_anomaly.

SeawaterPolynomials.TEOS10

SeawaterPolynomials.TEOS10.TEOS10SeawaterPolynomial — Type

struct TEOS10SeawaterPolynomial{FT} <: AbstractSeawaterPolynomial end

A 55-term polynomial approximation to the TEOS-10 standard equation of state for seawater.

SeawaterPolynomials.TEOS10.TEOS10SeawaterPolynomial — Type

TEOS10SeawaterPolynomial(FT=Float64)

Returns an object representing a 55-term polynomial approximation to the TEOS-10 standard equation of state for seawater.

See

Roquet et al., "Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard", Ocean Modelling (2015).

SeawaterPolynomials.TEOS10.TEOS10EquationOfState — Function

TEOS10EquationOfState(FT=Float64; reference_density=1020)

Returns an BoussinesqEquationOfState with a TEOS10SeawaterPolynomial of float type FT with reference density = 1020 kg m⁻³, the value used by

Roquet et al., "Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard", Ocean Modelling (2015).

when fitting polynomial coefficients to the full TEOS-10 standard equation of state. See the discussion prior to equation 8 in Roquet et al. (2015).

Note that according to Roquet et al. (2015):

"In a Boussinesq model, the choice of the ρ₀ value is important, yet it varies significantly among OGCMs, as it is a matter of personal preference."

SeawaterPolynomials.haline_contraction — Method

haline_contraction(Θ, Sᴬ, Z, ::TEOS10)

Returns the Boussinesq haline contraction coefficient $∂ρ/∂Sᴬ$ [kg/m³/(g/kg)] computed using the 55-term polynomial approximation to TEOS-10 described in Roquet et al. (3.1, 2014).

Inputs

Θ: conservative temperature (ITS-90) [°C]
Sᴬ: absolute salinity [g/kg]
Z: geopotential depth [m]

Output

β: Boussinesq haline contraction coefficient ∂ρ/∂Sᴬ [kg/m³/(g/kg)]

References

Roquet, F., Madec, G., McDougall, T. J., Barker, P. M., 2014: Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard. Ocean Modelling.

SeawaterPolynomials.thermal_expansion — Method

thermal_expansion(Θ, Sᴬ, Z, ::TEOS10)

Returns the Boussinesq thermal expansion coefficient $-∂ρ/∂Θ$ [kg/m³/K] computed using the 55-term polynomial approximation to TEOS-10 described in Roquet et al. (3.1, 2014).

Inputs

Θ: conservative temperature (ITS-90) [°C]
Sᴬ: absolute salinity [g/kg]
Z: geopotential depth [m]

Output

α: Boussinesq thermal expansion coefficient -∂ρ/∂Θ [kg/m³/K]

References

Roquet, F., Madec, G., McDougall, T. J., Barker, P. M., 2014: Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard. Ocean Modelling.

SeawaterPolynomials.ρ′ — Method

ρ′(Θ, Sᴬ, Z, ::BoussinesqEquationOfState{<:TEOS10EquationOfState})

Returns the in-situ density of seawater with state (Θ, Sᴬ, Z) using the 55-term polynomial approximation to TEOS-10 described in Roquet et al. (3.1, 2014).

Inputs

Θ: conservative temperature (ITS-90) [°C]
Sᴬ: absolute salinity [g/kg]
Z: geopotential depth [m]

Output

ρ: in-situ density [kg/m³]

References

Roquet, F., Madec, G., McDougall, T. J., Barker, P. M., 2014: Accurate polynomial expressions for the density and specific volume of seawater using the TEOS-10 standard. Ocean Modelling.

SeawaterPolynomials.SecondOrderSeawaterPolynomials

SeawaterPolynomials.SecondOrderSeawaterPolynomials.SecondOrderSeawaterPolynomial — Type

struct SecondOrderSeawaterPolynomial{FT} <: AbstractSeawaterPolynomial

Container of coefficients for a second-order polynomial function of absolute salinity Sᴬ, conservative temperature Θ, and geopotential depth Z for seawater density.

The coefficients have the form

\[R_ᵦᵪᵩ ,\]

where $β, χ, φ$ denote the order of the term to which the coefficent corresponds: $β$ is the polynomial order of absolute salinity, $Sᴬ$, $χ$ is the polynomial order of conservative temperature, $Θ$, and $φ$ is the order of geopotential height, $Z$.

For a SecondOrderSeawaterPolynomial, $β + χ + φ < 3$.

The coefficient $R₁₁₀$ arises in the seawater polynomial as

seawater_polynomial(Θ, Sᴬ, Z) = ⋯ + R₁₁₀ * Sᴬ * Θ + ⋯

SeawaterPolynomials.SecondOrderSeawaterPolynomials.CabbelingRoquetSeawaterPolynomial — Function

CabbelingRoquetSeawaterPolynomial(FT=Float64)

Parameters for a minimal equation of state that describes cabbeling, optimized for the 'current' oceanic temperature and salinity distribution.

For more information, type help?> RoquetSeawaterPolynomial.

SeawaterPolynomials.SecondOrderSeawaterPolynomials.CabbelingThermobaricityRoquetSeawaterPolynomial — Function

CabbelingThermobaricityRoquetSeawaterPolynomial(FT=Float64)

Parameters for a minimal equation of state that describes cabbeling and thermobaric effects on sewater density, optimized for the 'current' oceanic temperature and salinity distribution.

For more information, type help?> RoquetSeawaterPolynomial.

SeawaterPolynomials.SecondOrderSeawaterPolynomials.FreezingRoquetSeawaterPolynomial — Function

FreezingRoquetSeawaterPolynomial(FT=Float64)

Parameters for a minimal equation of state that describes seawater density near its freezing point, optimized for the 'current' oceanic temperature and salinity distribution.

For more information, type help?> RoquetSeawaterPolynomial.

SeawaterPolynomials.SecondOrderSeawaterPolynomials.LinearRoquetSeawaterPolynomial — Function

LinearRoquetSeawaterPolynomial(FT=Float64)

Parameters for a linear equation of state optimized for the 'current' oceanic temperature and salinity distribution.

For more information, type help?> RoquetSeawaterPolynomial.

SeawaterPolynomials.SecondOrderSeawaterPolynomials.RoquetEquationOfState — Function

RoquetEquationOfState([FT=Float64,] coefficient_set=:SecondOrder; reference_density=1024.6)

Returns an BoussinesqEquationOfState with a RoquetSeawaterPolynomial corresponding to coefficient_set and with reference density = 1024.6 kg m⁻³, the average density of seawater at the surface of the world ocean.

Type

help?> RoquetSeawaterPolynomial

for options for the coefficient_set. The optimzed coefficient sets for the RoquetSeawaterPolynomial are listed in Table 3 in

Roquet et al., "Defining a Simplified yet 'Realistic' Equation of State for Seawater", Journal of Physical Oceanography (2015).

SeawaterPolynomials.SecondOrderSeawaterPolynomials.RoquetSeawaterPolynomial — Function

RoquetSeawaterPolynomial([FT=Float64,] coefficient_set=:SecondOrder)

Returns a SecondOrderSeawaterPolynomial with coefficients optimized by

Roquet et al., "Defining a Simplified yet 'Realistic' Equation of State for Seawater", Journal of Physical Oceanography (2015).

The coefficient_set is a symbol or string that selects one of the "sets" of optimized second order coefficients.

Coefficient sets

- `:Linear`: a linear equation of state, `ρ = ρᵣ + R₁₀₀ * Θ + R₀₁₀ * Sᴬ`

- `:Cabbeling`: includes quadratic temperature term,
                `ρ = ρᵣ + R₁₀₀ * Θ + R₀₁₀ * Sᴬ + R₀₂₀ * Θ^2`

- `:CabbelingThermobaricity`: includes 'thermobaricity' term,
                               `ρ = ρᵣ + R₁₀₀ * Θ + R₀₁₀ * Sᴬ + R₀₂₀ * Θ^2 - R₀₁₁ * Θ * Z`

- `:Freezing`: same as `:cabbeling_thermobaricity` with modified constants to increase
               accuracy near freezing

- `:SecondOrder`: includes quadratic salinity, halibaricity, and thermohaline term,
                   `ρ = ρᵣ + R₁₀₀ * Θ + R₀₁₀ * Sᴬ + R₀₂₀ * Θ^2 - R₀₁₁ * T * Z`
                         + R₂₀₀ * Sᴬ^2 - R₁₀₁ * Sᴬ * Z + R₁₁₀ * Sᴬ * Θ`

The optimized coefficients are reported in Table 3 of Roquet et al., "Defining a Simplified yet 'Realistic' Equation of State for Seawater", Journal of Physical Oceanography (2015), and further discussed around equations (12)–(15). The optimization minimizes errors in estimated horizontal density gradient estiamted from climatological temperature and salinity distributions between the 5 simplified forms chosen by Roquet et. al and the full-fledged TEOS-10 equation of state.

SeawaterPolynomials.SecondOrderSeawaterPolynomials.SecondOrderRoquetSeawaterPolynomial — Function

SecondOrderRoquetSeawaterPolynomial(FT=Float64)

Parameters for a fully second-order equation of state for seawater, optimized for the 'current' oceanic temperature and salinity distribution.

For more information, type help?> RoquetSeawaterPolynomial.